Creating a commons folder to move common stuff to that
